more tests and take appart applyFilter
